[Feat] DB서버 테이블 생성 및 전처리된 데이터 삽입 #10 models 디렉토리에 DB 테이블 생성, 삽입 코드 추가함

+++ server/src/models/create_tables.sql
@@ -1,8 +1,8 @@
 CREATE TABLE station (
 	id INT AUTO_INCREMENT PRIMARY KEY,
 	station_name VARCHAR(10) NOT NULL,
-	prev_station_time TINYINT NOT NULL,
-	next_station_time TINYINT NOT NULL,
+	prev_station_time SMALLINT NOT NULL,
+	next_station_time SMALLINT NOT NULL,
 	station_info TEXT NOT NULL,
 	passenger_info TEXT NOT NULL
 );
